diplom - part3 - ok

+++ ewm-service/src/main/java/ru/practicum/ewm/comment/CommentServiceImpl.java
@@ -3,42 +3,44 @@ package ru.practicum.ewm.comment;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestParam;
 import ru.practicum.ewm.comment.dto.CommentDto;
 import ru.practicum.ewm.comment.dto.NewCommentDto;
 import ru.practicum.ewm.comment.dto.UpdatedCommentDto;
 import ru.practicum.ewm.comment.mapper.CommentMapper;
 import ru.practicum.ewm.comment.model.Comment;
-import ru.practicum.ewm.event.EventRepository;
 import ru.practicum.ewm.event.EventService;
 import ru.practicum.ewm.event.model.Event;
 import ru.practicum.ewm.exception.exceptions.EntityNotFoundException;
+import ru.practicum.ewm.exception.exceptions.IncorrectRequestException;
 import ru.practicum.ewm.exception.exceptions.UserNotCreatorThisCommentException;
-import ru.practicum.ewm.user.UserRepository;
 import ru.practicum.ewm.user.UserService;
 import ru.practicum.ewm.user.model.User;
+import ru.practicum.ewm.utility.Constants;
 
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
 import java.time.LocalDateTime;
 import java.util.List;
 
 @Slf4j
 @Service
 @RequiredArgsConstructor(onConstructor_ = @Autowired)
-public class CommentServiceImpl {
-    UserRepository userRepository;
-    UserService userService;
-    EventService eventService;
-    EventRepository eventRepository;
-    CommentRepository commentRepository;
-    CommentMapper commentMapper;
+public class CommentServiceImpl implements CommentService {
+    private final UserService userService;
+    private final EventService eventService;
+    private final CommentRepository commentRepository;
+    private final CommentMapper commentMapper;
 
+    @Override
     @Transactional
-    CommentDto addNewComment(Long userId, Long eventId, NewCommentDto newCommentDto) throws EntityNotFoundException {
+    public CommentDto addNewComment(Long userId, Long eventId, NewCommentDto newCommentDto)
+            throws EntityNotFoundException {
         User user = userService.getUserById(userId);
         Event event = eventService.getEventById(eventId);
 
@@ -46,75 +48,78 @@ public class CommentServiceImpl {
         comment.setAuthor(user);
         comment.setEvent(event);
 
-        log.error("Обновлен комментарий к Event c ID-{}: {}", eventId, comment);
+        log.info("Добавлен комментарий к Event c ID-{}: {}", eventId, comment);
         return commentMapper.CommentToDto(commentRepository.save(comment));
     }
 
 
+    @Override
     @Transactional
-    public CommentDto updateComment(Long userId, Long eventId, Long commentId, UpdatedCommentDto updatedCommentDto) throws EntityNotFoundException {
-        Comment comment = getComment(userId, eventId, commentId);
-
+    public CommentDto updateComment(Long userId, Long eventId, Long commentId, UpdatedCommentDto updatedCommentDto)
+            throws EntityNotFoundException {
+        User user = userService.getUserById(userId);
+        Event event = eventService.getEventById(eventId);
+        Comment comment;
+        comment = commentRepository.findByIdAndAuthorAndEvent(commentId, user, event);
+        if (comment == null) {
+            throw new EntityNotFoundException(Comment.class, eventId);
+        }
         comment.setText(updatedCommentDto.getText());
         comment.setEditedDate(LocalDateTime.now());
-        log.error("Updated comment to Event with ID-{}: {}", eventId, comment);
+        log.info("Updated comment to Event with ID-{}: {}", eventId, comment);
         return commentMapper.CommentToDto(commentRepository.save(comment));
 
     }
 
+    @Override
     @Transactional
-    public CommentDto deleteComment(Long userId, Long eventId, Long commentId) throws EntityNotFoundException, UserNotCreatorThisCommentException {
+    public CommentDto deleteComment(Long userId, Long eventId, Long commentId)
+            throws EntityNotFoundException, UserNotCreatorThisCommentException {
         User user = userService.getUserById(userId);
-        Comment comment = getComment(userId, eventId, commentId);
-
+        Comment comment = getCommentById(commentId);
+        if (!comment.getAuthor().getId().equals(user.getId())) {
+            throw new UserNotCreatorThisCommentException(commentId, userId);
+        }
         commentRepository.deleteById(commentId);
-        log.error("Deleted comment to Event with ID-{}: {}", eventId, comment);
+        log.info("Deleted comment to Event with ID-{}: {}", eventId, comment);
         return commentMapper.CommentToDto(comment);
     }
 
+    @Override
     @Transactional
-    public Comment deleteCommentByAdmin(Long commentId) throws EntityNotFoundException, UserNotCreatorThisCommentException {
+    public Comment deleteCommentByAdmin(Long commentId)
+            throws EntityNotFoundException {
         Comment comment = getCommentById(commentId);
         commentRepository.deleteById(commentId);
-        log.error("Admin deleted comment-{}", comment);
+        log.info("Admin deleted comment-{}", comment);
         return comment;
     }
 
-    @Transactional
-    public List<CommentDto> findAllByUserAndEvent(Long userId, Long eventId, int from, int size) throws EntityNotFoundException {
+    @Override
+    @Transactional(readOnly = true)
+    public List<CommentDto> findAllByUserAndEvent(Long userId, Long eventId, int from, int size)
+            throws EntityNotFoundException {
         Pageable pageable = PageRequest.of((from / size), size);
         User user = userService.getUserById(userId);
         Event event = eventService.getEventById(eventId);
 
         List<Comment> listComments = commentRepository.findAllByAuthorAndEventId(user, event, pageable);
-        log.error("Displayed comment list to Event with ID-{} from User with ID-{}", eventId, userId);
+        log.info("Displayed comment list to Event with ID-{} from User with ID-{}", eventId, userId);
         return commentMapper.toCommentListDto(listComments);
     }
 
-    @Transactional(readOnly = true)
-    private Comment getComment(Long userId, Long eventId, Long commentId) throws EntityNotFoundException {
-        User user = userService.getUserById(userId);
-        Event event = eventService.getEventById(eventId);
-        Comment comment;
-        try {
-            comment = commentRepository.findByIdAndAuthorAndEvent(commentId, user, event);
-        } catch (RuntimeException e) {
-            log.error("Generated EntityNotFoundException");
-            throw new EntityNotFoundException(Comment.class, eventId);
-        }
-        return comment;
-    }
-
+    @Override
     @Transactional(readOnly = true)
     public List<CommentDto> findAllByUser(Long userId, Integer from, Integer size) throws EntityNotFoundException {
         User user = userService.getUserById(userId);
         Pageable pageable = PageRequest.of((from / size), size);
 
         List<Comment> listComments = commentRepository.findAllByAuthor(user, pageable);
-        log.error("Displayed comment list from User with ID-{}", userId);
+        log.info("Displayed comment list from User with ID-{}", userId);
         return commentMapper.toCommentListDto(listComments);
     }
 
+    @Override
     @Transactional(readOnly = true)
     public CommentDto getCommentDtoById(Long userId, Long commentId)
             throws EntityNotFoundException, UserNotCreatorThisCommentException {
@@ -123,8 +128,92 @@ public class CommentServiceImpl {
         if (!comment.getAuthor().getId().equals(user.getId())) {
             throw new UserNotCreatorThisCommentException(comment.getId(), user.getId());
         }
-        log.error("Displayed comment with ID-{} from User with ID-{}", commentId, userId);
+        log.info("Displayed comment with ID-{} from User with ID-{}", commentId, userId);
         return commentMapper.CommentToDto(comment);
+    }
+
+    @Override
+    @Transactional(readOnly = true)
+    public CommentDto getCommentDtoById(Long commentId) throws EntityNotFoundException {
+        log.info("Returned comment with ID-{}", commentId);
+        return commentMapper.CommentToDto(getCommentById(commentId));
+    }
+
+    @Override
+    @Transactional(readOnly = true)
+    public List<CommentDto> findListCommentsDto(String text, List<Long> authors, List<Long> events,
+                                                String rangeStart, String rangeEnd, String sort,
+                                                Integer from, Integer size) throws IncorrectRequestException {
+        if (authors != null) {
+            if (authors.size() == 1 && authors.get(0) == 0) {
+                log.error("Generated IncorrectRequestException");
+                throw new IncorrectRequestException();
+            }
+        }
+        if (events != null) {
+            if (events.size() == 1 && events.get(0) == 0) {
+                log.error("Generated IncorrectRequestException");
+                throw new IncorrectRequestException();
+            }
+        }
+        LocalDateTime start;
+        LocalDateTime end;
+
+        if (rangeStart == null || rangeEnd == null) {
+            start = LocalDateTime.now().minusYears(10);
+            end = LocalDateTime.now().plusYears(10);
+        } else {
+            start = LocalDateTime.parse(rangeStart, Constants.TIME_FORMATTER);
+            end = LocalDateTime.parse(rangeEnd, Constants.TIME_FORMATTER);
+        }
+
+        Sort srt = Sort.unsorted();
+        if (sort != null) {
+            switch (sort) {
+                case "CREATED_DATE":
+                    srt = Sort.by(Sort.Direction.DESC, "createdDate");
+                    break;
+                case "EDITED_DATE":
+                    srt = Sort.by(Sort.Direction.DESC, "editedDate");
+                    break;
+            }
+        }
+        Pageable pageable = PageRequest.of(from / size, size, srt);
+
+        List<Comment> comments;
+
+        if (text == null) {
+            if (authors == null && events != null) {
+                comments = commentRepository.findCommentsByEvents
+                        (events, start, end, pageable);
+            } else if (events == null && authors != null) {
+                comments = commentRepository.findCommentsByAuthors
+                        (authors, start, end, pageable);
+            } else if (events == null) {
+                comments = commentRepository.findComments
+                        (start, end, pageable);
+            } else {
+                comments = commentRepository.findCommentsByAuthorsAndEvents
+                        (authors, events, start, end, pageable);
+            }
+        } else {
+            if (authors == null && events != null) {
+                comments = commentRepository.findCommentsByTextAndEvents
+                        (text, events, start, end, pageable);
+            } else if (events == null && authors != null) {
+                comments = commentRepository.findCommentsByTextAndAuthors
+                        (text, authors, start, end, pageable);
+            } else if (events == null) {
+                comments = commentRepository.findCommentsByText
+                        (text, start, end, pageable);
+            } else {
+                comments = commentRepository.findCommentsByTextAndAuthorsAndEvents
+                        (text, authors, events, start, end, pageable);
+            }
+        }
+        List<CommentDto> listCommentDto = commentMapper.toCommentListDto(comments);
+        log.info("Listing comments {} ", comments);
+        return listCommentDto;
     }
 
     @Transactional(readOnly = true)
